+++ pyaerocom/io/test/test_read_mscw_ctm.py
@@ -1,41 +1,42 @@
 import pytest
 import os
 from pyaerocom.conftest import lustre_unavail, testdata_unavail
-from pyaerocom.io.read_emep import ReadEMEP, ts_type_from_filename
+from pyaerocom.io import ReadMscwCtm
+from pyaerocom.io.read_mscw_ctm import ts_type_from_filename
 from pyaerocom.griddeddata import GriddedData
 from pyaerocom.colocation import colocate_gridded_gridded
 from pyaerocom.colocation import ColocatedData
 
 
 def test_read_emep():
-    r = ReadEMEP()
+    r = ReadMscwCtm()
     assert r.data_id == None
     assert r.vars_provided == []
     assert r.filepath == None
     assert r.years_avail == []
     with pytest.raises(FileNotFoundError):
-        ReadEMEP(data_dir='/invalid')
+        ReadMscwCtm(data_dir='/invalid')
     with pytest.raises(FileNotFoundError):
-        ReadEMEP(filepath='/invalid/test.nc')
+        ReadMscwCtm(filepath='/invalid/test.nc')
 
 
 @testdata_unavail
 def test_read_emep_read_var(path_emep):
     with pytest.raises(ValueError):
-        r = ReadEMEP()
+        r = ReadMscwCtm()
         r.read_var('od550aer')
     with pytest.raises(ValueError):
-        r = ReadEMEP()
+        r = ReadMscwCtm()
         r.read_var('od550aer')
     data_dir = path_emep['data_dir']
     with pytest.raises(FileNotFoundError):
-        r = ReadEMEP(data_dir=data_dir)
+        r = ReadMscwCtm(data_dir=data_dir)
         r.read_var('od550aer')
 
 @testdata_unavail
 def test_read_emep_data(path_emep):
     path = path_emep['daily']
-    r = ReadEMEP(filepath=path)
+    r = ReadMscwCtm(filepath=path)
 
     vars_provided = r.vars_provided
     assert isinstance(vars_provided, list)
@@ -49,7 +50,7 @@ def test_read_emep_data(path_emep):
 @testdata_unavail
 def test_read_emep_directory(path_emep):
     data_dir = path_emep['data_dir']
-    r = ReadEMEP(data_dir=data_dir)
+    r = ReadMscwCtm(data_dir=data_dir)
     assert r.data_dir == data_dir
     vars_provided = r.vars_provided
     assert 'vmro3' in vars_provided
@@ -64,7 +65,7 @@ def test_read_emep_directory(path_emep):
 def test_read_emep_ts_types(files, ts_types, tmpdir):
     for filename in files:
         open(os.path.join(tmpdir, filename), 'a').close()
-    r = ReadEMEP(data_dir=tmpdir)
+    r = ReadMscwCtm(data_dir=tmpdir)
     assert sorted(r.ts_types) == sorted(ts_types)
 
 @pytest.mark.parametrize('filename,ts_type', [
@@ -78,17 +79,17 @@ def test_ts_type_from_filename(filename, ts_type):
 
 def test_read_emep_years_avail(path_emep):
     data_dir = path_emep['data_dir']
-    r = ReadEMEP(data_dir=data_dir)
+    r = ReadMscwCtm(data_dir=data_dir)
     assert r.years_avail == [2017]
 
 def test_preprocess_units():
     units = ''
     prefix = 'AOD'
-    assert ReadEMEP().preprocess_units(units, prefix) == '1'
+    assert ReadMscwCtm().preprocess_units(units, prefix) == '1'
 
     units = 'mgS/m2'
     with pytest.raises(NotImplementedError):
-        ReadEMEP().preprocess_units(units)
+        ReadMscwCtm().preprocess_units(units)
 
 if __name__ == '__main__':
     import sys
